rename the mplex package to mplex
