+++ p2p/muxer/mplex/stream.go
@@ -1,4 +1,4 @@
-package peerstream_multiplex
+package mplex
 
 import (
 	"time"
