+++ p2p/muxer/mplex/transport.go
@@ -1,4 +1,4 @@
-package peerstream_multiplex
+package mplex
 
 import (
 	"net"
