+++ p2p/muxer/mplex/transport_test.go
@@ -1,4 +1,4 @@
-package peerstream_multiplex
+package mplex
 
 import (
 	"errors"
